Refactor/remove ruby21 patch (#1873)
* Remove ruby 2.1 patch

* Moved top level errors out of runtime to bespoke errors file

* Re-order errors according to names

* Re-order requires and pivot to more optimal require relatives

* Tidy up of rubocop autogen file
