+++ .rubocop_todo.yml
@@ -1,16 +1,15 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2026-03-17 23:57:03 UTC using RuboCop version 1.85.1.
+# on 2026-05-15 19:52:05 UTC using RuboCop version 1.85.1.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
 # versions of RuboCop, may require this file to be generated again.
 
-# TODO - [LH] -> Jan '25 (Updated deps and v10 prep) - 369 files inspected, 704 offenses detected, 112 offenses autocorrectable
-# TODO - [LH] -> Mar '25 (v10 prep) - 370 files inspected, 721 offenses detected, 116 offenses autocorrectable
 # TODO - [LH] -> Dec '25 - 375 files inspected, 713 offenses detected, 109 offenses autocorrectable
 # TODO - [LH] -> Dec '25 (query prep) - 378 files inspected, 729 offenses detected, 109 offenses autocorrectable
 # TODO - [LH] -> Mar '26 (v11 prep) - 389 files inspected, 747 offenses detected, 108 offenses autocorrectable
+# TODO - [LH] -> May '26 (v11.0.0) - 391 files inspected, 740 offenses detected, 105 offenses autocorrectable
 
 # Offense count: 1
 # This cop supports safe autocorrection (--autocorrect).
@@ -18,6 +17,20 @@
 Layout/EmptyLineBetweenDefs:
   Exclude:
     - 'lib/cucumber/glue/snippet.rb'
+
+# Offense count: 1
+# This cop supports safe autocorrection (--autocorrect).
+# Configuration parameters: EnforcedStyle.
+# SupportedStyles: empty_lines, empty_lines_except_namespace, empty_lines_special, no_empty_lines
+Layout/EmptyLinesAroundModuleBody:
+  Exclude:
+    - 'lib/cucumber/glue/multiple_world.rb'
+
+# Offense count: 1
+# This cop supports safe autocorrection (--autocorrect).
+Layout/HeredocIndentation:
+  Exclude:
+    - 'lib/cucumber/glue/multiple_world.rb'
 
 # Offense count: 2
 Lint/IneffectiveAccessModifier:
@@ -39,12 +52,6 @@ Lint/UselessAccessModifier:
   Exclude:
     - 'lib/cucumber/formatter/curl_option_parser.rb'
 
-# Offense count: 1
-# This cop supports safe autocorrection (--autocorrect).
-Lint/UselessAssignment:
-  Exclude:
-    - 'compatibility/cck_spec.rb'
-
 # Offense count: 2
 # This cop supports unsafe autocorrection (--autocorrect-all).
 Lint/UselessMethodDefinition:
@@ -65,7 +72,7 @@ Metrics/BlockLength:
 # Offense count: 14
 # Configuration parameters: CountComments, CountAsOne.
 Metrics/ClassLength:
-  Max: 516
+  Max: 500
 
 # Offense count: 8
 # Configuration parameters: AllowedMethods, AllowedPatterns.
@@ -77,7 +84,7 @@ Metrics/CyclomaticComplexity:
 Metrics/MethodLength:
   Max: 65
 
-# Offense count: 12
+# Offense count: 11
 # Configuration parameters: CountComments, CountAsOne.
 Metrics/ModuleLength:
   Max: 804
@@ -150,7 +157,7 @@ RSpec/EmptyExampleGroup:
     - 'spec/cucumber/filters/activate_steps_spec.rb'
     - 'spec/cucumber/running_test_case_spec.rb'
 
-# Offense count: 70
+# Offense count: 69
 # Configuration parameters: CountAsOne.
 RSpec/ExampleLength:
   Max: 58
@@ -187,7 +194,7 @@ RSpec/ExpectOutput:
   Exclude:
     - 'spec/cucumber/formatter/interceptor_spec.rb'
 
-# Offense count: 58
+# Offense count: 54
 # This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: EnforcedStyle.
 # SupportedStyles: implicit, each, example
@@ -211,12 +218,6 @@ RSpec/InstanceVariable:
     - 'spec/cucumber/formatter/query/hook_by_test_step_spec.rb'
     - 'spec/support/shared_context/http_server.rb'
 
-# Offense count: 1
-# This cop supports safe autocorrection (--autocorrect).
-RSpec/LeadingSubject:
-  Exclude:
-    - 'spec/cucumber/cli/main_spec.rb'
-
 # Offense count: 2
 # This cop supports safe autocorrection (--autocorrect).
 RSpec/MatchArray:
@@ -238,7 +239,7 @@ RSpec/MissingExampleGroupArgument:
     - 'spec/cucumber/formatter/fail_fast_spec.rb'
     - 'spec/cucumber/formatter/rerun_spec.rb'
 
-# Offense count: 58
+# Offense count: 57
 RSpec/MultipleExpectations:
   Max: 3
 
@@ -261,7 +262,7 @@ RSpec/NamedSubject:
     - 'spec/cucumber/cli/main_spec.rb'
     - 'spec/cucumber/configuration_spec.rb'
     - 'spec/cucumber/formatter/url_reporter_spec.rb'
-    - 'spec/cucumber/rake/forked_spec.rb'
+    - 'spec/cucumber/rake/forked_cucumber_runner_spec.rb'
     - 'spec/cucumber/rake/task_spec.rb'
     - 'spec/cucumber/runtime/meta_message_builder_spec.rb'
     - 'spec/cucumber/runtime/support_code_spec.rb'
@@ -287,7 +288,7 @@ RSpec/OverwritingSetup:
   Exclude:
     - 'spec/cucumber/runtime/support_code_spec.rb'
 
-# Offense count: 10
+# Offense count: 11
 # This cop supports unsafe autocorrection (--autocorrect-all).
 RSpec/ReceiveMessages:
   Exclude:
@@ -329,7 +330,7 @@ RSpec/SubjectDeclaration:
     - 'spec/cucumber/runtime/after_hooks_spec.rb'
     - 'spec/cucumber/runtime/before_hooks_spec.rb'
 
-# Offense count: 53
+# Offense count: 52
 # Configuration parameters: IgnoreNameless, IgnoreSymbolicNames.
 RSpec/VerifiedDoubles:
   Exclude:
