+++ lib/cucumber/errors.rb
@@ -3,6 +3,49 @@
 require 'cucumber/core/test/result'
 
 module Cucumber
+  # Raised when a step matches 2 or more StepDefinitions
+  class Ambiguous < StandardError
+    def initialize(step_name, step_definitions, used_guess)
+      # TODO: [LH] - Just use a heredoc here to fix this up
+      message = String.new
+      message << "Ambiguous match of \"#{step_name}\":\n\n"
+      message << step_definitions.map(&:backtrace_line).join("\n")
+      message << "\n\n"
+      message << "You can run again with --guess to make Cucumber be more smart about it\n" unless used_guess
+      super(message)
+    end
+  end
+
+  class FeatureFolderNotFoundException < RuntimeError
+    def initialize(path)
+      @path = path
+      super
+    end
+
+    def message
+      "No such file or directory - #{@path}"
+    end
+  end
+
+  class FileNotFoundException < RuntimeError
+    attr_reader :path
+
+    def initialize(original_exception, path)
+      @path = path
+      super(original_exception)
+    end
+  end
+
+  # Raised when a StepDefinition's block invokes World#pending
+  class Pending < Core::Test::Result::Pending
+  end
+
+  class TagExcess < StandardError
+    def initialize(messages)
+      super(messages.join("\n"))
+    end
+  end
+
   # Raised when there is no matching StepDefinition for a step.
   class Undefined < Core::Test::Result::Undefined
     def self.from(result, step_name)
@@ -25,29 +68,6 @@ module Cucumber
   class UndefinedDynamicStep < StandardError
     def initialize(step_name)
       super %(Undefined dynamic step: "#{step_name}")
-    end
-  end
-
-  # Raised when a StepDefinition's block invokes World#pending
-  class Pending < Core::Test::Result::Pending
-  end
-
-  # Raised when a step matches 2 or more StepDefinitions
-  class Ambiguous < StandardError
-    def initialize(step_name, step_definitions, used_guess)
-      # TODO: [LH] - Just use a heredoc here to fix this up
-      message = String.new
-      message << "Ambiguous match of \"#{step_name}\":\n\n"
-      message << step_definitions.map(&:backtrace_line).join("\n")
-      message << "\n\n"
-      message << "You can run again with --guess to make Cucumber be more smart about it\n" unless used_guess
-      super(message)
-    end
-  end
-
-  class TagExcess < StandardError
-    def initialize(messages)
-      super(messages.join("\n"))
     end
   end
 end
